Replace setup scripts with custom image
Instead of interfacing with `toolbox` using a setup script,
we now build custom images using a `Containerfile`. We base
our images directly off `registry.fedoraproject.org/fedora-toolbox:$ver`,
where `$ver` is the Fedora version. These images already do
a lot of nifty things [1], in order to ensure that everything
works as expected. Therefore it is highly encouraged to
built on top of that existing infrastructure.

[1] https://github.com/containers/toolbox/blob/main/images/fedora/f39/Containerfile