Add new installation method for slint tools
Previously we used cargo install in order to install the
slint lsp into our development container. The downside to
that approach is that this takes a lot of time, as it compiles
the binary from source. What we do now is download the latest
version of slint lsp and slint viewer from the GitHub releases,
extract it and put it into the /usr/bin directory.