Change deployment mechanism of sandbox
Do not patch /etc/profile mindlessly because this can cause toolbox run
to fail. Instead we mandate the user to use the toolbox run command in
order to enter a sandboxed environment. This way we can also modify the
container on the fly by running toolbox enter instead of toolbox run
sandbox.